+++ RELEASE_NOTES.md
@@ -3,11 +3,26 @@
 
 ## 0.2.0
 
-This release tweaks the behavior of the bottom sheet.
+This release tweaks the design and the behavior of the bottom sheet.
+
+The handle was previously placed in a non-designable area, under which the sheet content scrolled.
+
+Now, the area has a divider that separates the handle from the content, with the entire handle bar area being customizable with the handle style. 
+
+### 💥 New Features
+
+* There is a new `BottomSheetHandleBar` view.
+* `BottomSheetHandleStyle` has a new `backgroundColor` property.
+* `BottomSheetHandleStyle` has a new `dividerColor` property.
+* `BottomSheetHandleStyle` has a new `padding` property.
 
 ### 💡 Behavior Changes
 
-* The sheet handle is placed above the content using a ZStack, instead of a VStack. This avoids content being clipped.
+* The sheet handle is now placed in a bar, with padding and a divider.
+
+### 🗑 Deprecations
+
+* `BottomSheetHandleStyle` `color` has been renamed to `handleColor`.
 
 
 ## 0.1.0